+++ Test.Crawler/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using Komodo.Classes;
 using Komodo.Crawler;
+using BlobHelper;
 
 namespace Test.Crawler
 {
@@ -13,11 +14,20 @@ namespace Test.Crawler
             while (true)
             {
                 Console.WriteLine("Press ENTER to quit");
-                string crawlerType = Common.InputString("Crawler type [file|ftp|http|sql|sqlite]:", "file", true);
+                string crawlerType = Common.InputString("Crawler type [s3|azure|kvp|file|ftp|http|sql|sqlite]:", "file", true);
                 if (String.IsNullOrEmpty(crawlerType)) break;
 
                 switch (crawlerType)
                 {
+                    case "s3":
+                        S3Crawler();
+                        break;
+                    case "azure":
+                        AzureCrawler();
+                        break;
+                    case "kvp":
+                        KvpCrawler();
+                        break;
                     case "file":
                         FileCrawler();
                         break;
@@ -37,6 +47,72 @@ namespace Test.Crawler
                         break;
                 }
             }
+        }
+
+        static void S3Crawler()
+        { 
+            string endpoint = Common.InputString("Endpoint:", null, true);
+            bool ssl = Common.InputBoolean("SSL:", true);
+            string bucket = Common.InputString("Bucket:", null, false);
+            string key = Common.InputString("Key:", null, false);
+            string accessKey = Common.InputString("Access Key:", null, false);
+            string secretKey = Common.InputString("Secret Key:", null, false);
+            AwsRegion region = (AwsRegion)(Enum.Parse(typeof(AwsRegion), Common.InputString("Region:", "USWest1", false)));
+            string baseUrl = Common.InputString("Base URL:", "http://localhost:8000/{bucket}/{key}", false);
+
+            S3Crawler s3c = null;
+            if (!String.IsNullOrEmpty(endpoint)) s3c = new S3Crawler(endpoint, ssl, bucket, key, accessKey, secretKey, region, baseUrl);
+            else s3c = new S3Crawler(bucket, key, accessKey, secretKey, region);
+            
+            S3CrawlResult s3cr = s3c.Get();
+
+            Console.WriteLine("Success        : " + s3cr.Success);
+            Console.WriteLine("Start time     : " + s3cr.Time.Start.ToString());
+            Console.WriteLine("End time       : " + s3cr.Time.End.ToString());
+            Console.WriteLine("Total ms       : " + s3cr.Time.TotalMs.ToString() + "ms");
+            Console.WriteLine("Content length : " + s3cr.ContentLength + " bytes"); 
+            Console.WriteLine("Data           :" + Environment.NewLine + Encoding.UTF8.GetString(s3cr.Data));
+        }
+
+        static void AzureCrawler()
+        { 
+            string accountName = Common.InputString("Account Name:", null, true); 
+            string container = Common.InputString("Container:", null, false);
+            string endpoint = Common.InputString("Endpoint:", null, false);
+            string accessKey = Common.InputString("Access Key:", null, false);
+            string key = Common.InputString("Key:", null, false);
+
+            AzureBlobCrawler ac = new AzureBlobCrawler(accountName, container, endpoint, accessKey, key);
+
+            AzureBlobCrawlResult ar = ac.Get();
+
+            Console.WriteLine("Success        : " + ar.Success);
+            Console.WriteLine("Start time     : " + ar.Time.Start.ToString());
+            Console.WriteLine("End time       : " + ar.Time.End.ToString());
+            Console.WriteLine("Total ms       : " + ar.Time.TotalMs.ToString() + "ms");
+            Console.WriteLine("Content length : " + ar.ContentLength + " bytes");
+            Console.WriteLine("Data           :" + Environment.NewLine + Encoding.UTF8.GetString(ar.Data));
+        }
+
+        static void KvpCrawler()
+        {
+            // string endpoint, string userGuid, string container, string apiKey, string key
+            string endpoint = Common.InputString("Endpoint:", null, false);
+            string userGuid = Common.InputString("User GUID:", null, true);
+            string container = Common.InputString("Container:", null, false);
+            string apiKey = Common.InputString("API Key:", null, false);
+            string key = Common.InputString("Key:", null, false);
+
+            KvpbaseCrawler kc = new KvpbaseCrawler(endpoint, userGuid, container, apiKey, key);
+
+            KvpbaseCrawlResult kcr = kc.Get();
+
+            Console.WriteLine("Success        : " + kcr.Success);
+            Console.WriteLine("Start time     : " + kcr.Time.Start.ToString());
+            Console.WriteLine("End time       : " + kcr.Time.End.ToString());
+            Console.WriteLine("Total ms       : " + kcr.Time.TotalMs.ToString() + "ms");
+            Console.WriteLine("Content length : " + kcr.ContentLength + " bytes");
+            Console.WriteLine("Data           :" + Environment.NewLine + Encoding.UTF8.GetString(kcr.Data));
         }
 
         static void FileCrawler()